+++ stac-api/src/item_collection.rs
@@ -12,6 +12,7 @@ const ITEM_COLLECTION_TYPE: &str = "FeatureCollection";
 /// not be. Defined by the [itemcollection
 /// fragment](https://github.com/radiantearth/stac-api-spec/blob/main/fragments/itemcollection/README.md).
 #[derive(Debug, Serialize, Deserialize)]
+#[cfg_attr(feature = "schemars", derive(schemars::JsonSchema))]
 pub struct ItemCollection {
     /// Always "FeatureCollection" to provide compatibility with GeoJSON.
     pub r#type: String,
@@ -46,6 +47,7 @@ pub struct ItemCollection {
 ///
 /// Part of the [context extension](https://github.com/stac-api-extensions/context).
 #[derive(Debug, Serialize, Deserialize)]
+#[cfg_attr(feature = "schemars", derive(schemars::JsonSchema))]
 pub struct Context {
     /// The count of results returned by this response. Equal to the cardinality
     /// of features array.