+++ stac-api/src/sort.rs
@@ -7,6 +7,7 @@ use std::{
 
 /// Fields by which to sort results.
 #[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
+#[cfg_attr(feature = "schemars", derive(schemars::JsonSchema))]
 pub struct Sortby {
     /// The field to sort by.
     pub field: String,
@@ -16,6 +17,7 @@ pub struct Sortby {
 }
 
 #[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
+#[cfg_attr(feature = "schemars", derive(schemars::JsonSchema))]
 pub enum Direction {
     #[serde(rename = "asc")]
     Ascending,