+++ stac-api/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Conformance URIs ([#170](https://github.com/gadomski/stac-rs/pull/170))
+- `schemars` feature ([#177](https://github.com/gadomski/stac-rs/pull/177))
 
 ### Changed